requirements.txt
# Core data stack
numpy>=1.26,<2.0
pandas>=2.2,<3.0
PyYAML>=6.0,<7.0

# Graph and ML stack
torch>=2.2,<3.0
torch-geometric>=2.5,<3.0
scikit-learn>=1.5,<2.0
xgboost>=2.1,<3.0
networkx>=3.2,<4.0
# Optional accelerator:
# `pymetis` is intentionally excluded from the base install because it often
# requires a native METIS toolchain and fails on Windows when Meson picks a
# 32-bit MinGW compiler ahead of 64-bit Python. The runtime partitioner already
# falls back to a deterministic implementation when `pymetis` is unavailable.

# API and infrastructure
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.8,<3.0
redis>=5.0,<6.0
boto3>=1.35,<2.0
httpx>=0.28,<1.0
prometheus-client>=0.21,<1.0

# Test tooling
pytest>=9.0,<10.0
